refactor(Services.Teaser+Unified): #402c2 drop import Search via factory move to ServiceContainer
Drops `import Search` from both Services.TeaserService.swift and
Services.UnifiedSearchService.swift by moving their static
`with*Service` factories to Services.ServiceContainer.swift —
ServiceContainer keeps its `import Search` for the Search.Index
instantiation responsibility, but the two service files become Search-
free.

Per-file changes:

1. Services/ReadCommands/Services.TeaserService.swift:
   - Replace `init(searchDbPath: URL?, sampleDbPath: URL?)` with
     `init(searchIndex: (any Search.Database)?, sampleDbPath: URL?)`.
     Caller passes the (already-constructed) Search.Database directly;
     this file no longer touches Search.Index(dbPath:).
   - Remove the `extension Services.ServiceContainer { withTeaserService }`
     block at the bottom of the file. The factory now lives in
     Services.ServiceContainer.swift alongside withDocsService /
     withHIGService / withSampleService.
   - Drop `import Search`.

2. Services/ReadCommands/Services.UnifiedSearchService.swift:
   - Same pattern: replace `init(searchDbPath:, sampleDbPath:)` with
     `init(searchIndex:, sampleDbPath:)`.
   - Remove the `extension Services.ServiceContainer { withUnifiedSearchService }`
     block.
   - Drop `import Search`.

3. Services/Services.ServiceContainer.swift gains two new factory
   methods (lifted from the per-service files):

       public static func withTeaserService<T: Sendable>(
           searchDbPath: String? = nil,
           sampleDbPath: URL? = nil,
           operation: (Services.TeaserService) async throws -> T,
       ) async throws -> T { ... }

       public static func withUnifiedSearchService<T: Sendable>(
           searchDbPath: String? = nil,
           sampleDbPath: URL? = nil,
           operation: (Services.UnifiedSearchService) async throws -> T,
       ) async throws -> T { ... }

   Both methods resolve the search-DB + sample-DB paths via the
   existing Shared.Utils.PathResolver helpers, optionally instantiate
   `Search.Index(dbPath:)`, and pass the resulting `any Search.Database?`
   to the protocol-typed init. CLI / MCP / TUI callers using these
   `with*Service` static methods continue to compile unchanged —
   the API surface seen by consumers is identical.

After this PR, Services' import-Search file count drops from 4 to 2.
The remaining importers are ServiceContainer.swift (the composition
root inside Services that constructs Search.Index) and ReadService.swift
(uses Search.PackageQuery, a separate seam — slice C5).

Verification: xcrun swift build clean; make test-clean -> 1414 tests
in 157 suites pass.

Part 9 of #402.
